Consolidate unit tests into unit test class
