+++ platforms/jvm/plugins-java/src/test/groovy/org/gradle/api/plugins/JavaPluginTest.groovy
@@ -17,6 +17,7 @@
 package org.gradle.api.plugins
 
 import org.gradle.api.DefaultTask
+import org.gradle.api.JavaVersion
 import org.gradle.api.artifacts.Dependency
 import org.gradle.api.component.AdhocComponentWithVariants
 import org.gradle.api.internal.component.SoftwareComponentInternal
@@ -27,6 +28,9 @@ import org.gradle.api.tasks.SourceSet
 import org.gradle.api.tasks.bundling.Jar
 import org.gradle.api.tasks.compile.JavaCompile
 import org.gradle.api.tasks.javadoc.Javadoc
+import org.gradle.api.tasks.testing.Test
+import org.gradle.internal.jvm.Jvm
+import org.gradle.jvm.toolchain.JavaLanguageVersion
 import org.gradle.test.fixtures.AbstractProjectBuilderSpec
 import org.gradle.test.fixtures.file.TestFile
 import org.gradle.util.TestUtil
@@ -527,6 +531,75 @@ class JavaPluginTest extends AbstractProjectBuilderSpec {
         task.taskDependencies.getDependencies(task)*.path as Set == [':middle:build', ':app:buildDependents'] as Set
     }
 
+    def "wires toolchain for sourceset if toolchain is configured"() {
+        given:
+        def someJdk = Jvm.current()
+        setupProjectWithToolchain(someJdk.javaVersion)
+
+        when:
+        project.sourceSets.create('custom')
+
+        then:
+        def compileTask = project.tasks.named("compileCustomJava", JavaCompile).get()
+        def configuredToolchain = compileTask.javaCompiler.get().javaToolchain
+        configuredToolchain.displayName.contains(someJdk.javaVersion.getMajorVersion())
+    }
+
+    def "source and target compatibility are configured if toolchain is configured"() {
+        given:
+        setupProjectWithToolchain(Jvm.current().javaVersion)
+
+        when:
+        project.sourceSets.create('custom')
+
+        then:
+        JavaVersion.toVersion(project.tasks.compileJava.getSourceCompatibility()).majorVersion == Jvm.current().javaVersion.majorVersion
+        JavaVersion.toVersion(project.tasks.compileJava.getTargetCompatibility()).majorVersion == Jvm.current().javaVersion.majorVersion
+        JavaVersion.toVersion(project.tasks.compileCustomJava.getSourceCompatibility()).majorVersion == Jvm.current().javaVersion.majorVersion
+        JavaVersion.toVersion(project.tasks.compileCustomJava.getTargetCompatibility()).majorVersion == Jvm.current().javaVersion.majorVersion
+    }
+
+    def "wires toolchain for test if toolchain is configured"() {
+        given:
+        def someJdk = Jvm.current()
+        setupProjectWithToolchain(someJdk.javaVersion)
+
+        when:
+        def testTask = project.tasks.named("test", Test).get()
+        def configuredJavaLauncher = testTask.javaLauncher.get()
+
+        then:
+        configuredJavaLauncher.executablePath.toString().contains(someJdk.javaVersion.getMajorVersion())
+    }
+
+    def "wires toolchain for javadoc if toolchain is configured"() {
+        given:
+        def someJdk = Jvm.current()
+        setupProjectWithToolchain(someJdk.javaVersion)
+
+        when:
+        def javadocTask = project.tasks.named("javadoc", Javadoc).get()
+        def configuredJavadocTool = javadocTask.javadocTool
+
+        then:
+        configuredJavadocTool.isPresent()
+    }
+
+    def 'can set java compile source compatibility if toolchain is configured'() {
+        given:
+        def someJdk = Jvm.current()
+        setupProjectWithToolchain(someJdk.javaVersion)
+        def prevJavaVersion = JavaVersion.toVersion(someJdk.javaVersion.majorVersion.toInteger() - 1)
+        project.java.sourceCompatibility = prevJavaVersion
+
+        when:
+        def javaCompileTask = project.tasks.named("compileJava", JavaCompile).get()
+
+        then:
+        javaCompileTask.sourceCompatibility == prevJavaVersion.toString()
+        javaCompileTask.sourceCompatibility == prevJavaVersion.toString()
+    }
+
     void "source and target compatibility of compile tasks default to release if set"() {
         given:
         project.pluginManager.apply(JavaPlugin)
@@ -542,5 +615,10 @@ class JavaPluginTest extends AbstractProjectBuilderSpec {
         compileJava.sourceCompatibility == "1.8"
         testCompileJava.targetCompatibility == "9"
         testCompileJava.sourceCompatibility == "9"
+    }
+
+    private void setupProjectWithToolchain(JavaVersion version) {
+        project.pluginManager.apply(JavaPlugin)
+        project.java.toolchain.languageVersion = JavaLanguageVersion.of(version.majorVersion)
     }
 }
